Remake controller entirely to be smaller and work
The old controller was way too big and did not have the sides recognised
by CURA and the 3D printer. The proportions of the new controller are
better and also have a better design and proper, planned measurements.
I did this using an actual plan on paper which can be seen in the
/photos/ directory.
Moved the /photos/ directory to the project root.
Also added a file rundown of all the files in README.md.