Fix movement and check for wall collision
Before, the sprite was being drawn at an offset and rotated incorrectly
due to the rotation being set around the bottom-right corner instead of
the center of the sprite.
Wall collision is smart and checks for transparency; the transparent
parts of the sprite can go through black walls.

Also made Stage 3 simpler. More stages in increasing difficulty will
be added later.