Build and program the controller
The controller has the following inputs: ON/OFF, A, B, X, Y, START,
Joystick X and Y, and Joystick Z (pushing down the joystick).

The Communicator class now works asynchronouysly/as a separate thread so
that it can scan the controller's inputs without lagging the game.
Also made the player model slightly smaller to make the game easier with
a controller.